Extract thread lifecycle integration tests
