+++ docs/DECISIONS.md
@@ -264,3 +264,4 @@
 - Composer integration coverage lives in `WorkspaceComposerIntegrationTests`, not `WorkspaceModelTests`. Composer submit, tool-card creation, artifact surfacing, Computer Use dispatch, queued-tool progress, streaming assistant drafts, cancellation notices, empty-draft no-ops, and selection races cross the workspace model, agent runner, transcript events, tool cards, and top-bar status; keep those flows together while pure submission/cancellation planning stays in focused planner tests.
 - Message feedback and artifact preview coverage should stay out of `WorkspaceModelTests`. Feedback persistence lives in `WorkspaceFeedbackIntegrationTests` because it crosses model mutation, thread events, and transcript surfaces. Pure `ToolArtifactState` image/document/text preview derivation lives in `QuillCodeToolCardSurfaceTests` because it is value-type surface behavior independent of workspace state.
 - Runtime issue integration coverage lives in `WorkspaceRuntimeIssueIntegrationTests`, not `WorkspaceModelTests`. Applying runtime status, projecting runtime issues into top-bar/settings surfaces, diagnostic redaction, and retry-last-turn mutation cross runtime state, workspace surfaces, commands, and composer state; keep pure issue construction in `WorkspaceRuntimeIssueBuilderTests` and recovery-action mapping in `QuillCodeRuntimeIssueRecoveryPlannerTests`.
+- Thread lifecycle integration coverage lives in `WorkspaceThreadLifecycleIntegrationTests`, not `WorkspaceModelTests`. New chat selection, fork/compact context, project-thread fallback selection, pin/archive persistence, and rename/duplicate/unarchive/delete flows cross the workspace model, stores, sidebar ordering, selected project, and top-bar state; keep pure mutation algorithms in `WorkspaceThreadLifecycleEngineTests`.
